+++ apps/backend/src/modules/vendors/libs/enums/enums.ts
@@ -3,7 +3,6 @@ export { EdxCourseField } from "./edx-course-field.enum.js";
 export { UdemyApiPath } from "./udemy-api-path.enum.js";
 export { CourseField } from "./udemy-course-field.enum.js";
 export { CourseSectionField } from "./udemy-course-section-field.enum.js";
-export { UdemyDefaultSearchPageParameter } from "./udemy-default-search-page-parameter.enum.js";
 export { UdemyPageParameter } from "./udemy-page-parameter.enum.js";
 export {
 	VendorErrorMessage,