+++ apps/frontend/src/modules/courses/courses.ts
@@ -18,5 +18,6 @@ export {
 export {
 	type CourseDto,
 	type CourseSearchFilterDto,
+	type CourseSearchResponseDto,
 } from "./libs/types/types.js";
 export { actions, reducer } from "./slices/courses.js";