+++ apps/frontend/src/pages/feed/libs/components/feed-activity/feed-activity.tsx
@@ -60,9 +60,18 @@ const FeedActivity: React.FC<Properties> = ({
 			</div>
 			<div className={styles["card-content"]}>
 				<div>
-					<h3 className={styles["activity-title"]}>
-						{activity.user.firstName} {activity.user.lastName}
-					</h3>
+					<div className={styles["activity-title-container"]}>
+						<Link to={avatarLink as ValueOf<typeof AppRoute>}>
+							<Image
+								alt="User avatar"
+								className={styles["small-photo"]}
+								src={activity.user.avatarUrl ?? defaultAvatar}
+							/>
+						</Link>
+						<h3 className={styles["activity-title"]}>
+							{activity.user.firstName} {activity.user.lastName}
+						</h3>
+					</div>
 					<p className={styles["card-info"]}>
 						{getActivityTitle(activity, userId)}
 					</p>